main refactoring2
