+++ example/agents/agentMAB.py
@@ -1,3 +1,4 @@
+import os
 import gym
 import numpy as np
 
@@ -19,6 +20,13 @@ for i in range(a_size):
 
 mean = np.zeros(8)
 maxreward = 1
+
+dir = './results/agent_mab/'
+if not os.path.exists(dir):
+    os.makedirs(dir)
+logfile = dir + 'raw.csv'
+with open(logfile, 'w') as fd:
+    fd.write("\n")
 
 while True:
     print("--------------------------------------------------------")
@@ -43,7 +51,7 @@ while True:
     avg[action] = (avg[action] * num[action] + reward) / (num[action] +1)
     num[action] += 1
     
-    with open('result.csv','a') as fd:
+    with open(logfile,'a') as fd:
         fd.write(str(action) + "," + str(reward) + "\n")
     
     print("Mean:", str(avg))
@@ -51,4 +59,5 @@ while True:
     
     if done:
         break
+
 env.close()
