+++ example/agents/agentRL.py
@@ -8,7 +8,23 @@ from keras.models import Sequential
 from keras.layers import Dense
 from keras.optimizers import Adam
 
-from utils import float_to_bin_vec, bin_vec_to_float
+def float_to_bin_vec(v, no_bits=10):
+    scale = np.power(2, no_bits-1)
+
+    # convert to binary vector
+    y = list(bin(round(v * scale))[2:])
+    y = [int(v) for v in y]
+    y2 = list(np.zeros(no_bits - len(y)))
+    y2.extend(y)
+
+    return np.reshape(y2, [1, no_bits])
+
+def bin_vec_to_float(b, no_bits=10):
+    p = [np.power(2, x) for x in range(no_bits-1, -1, -1)]
+    scale = np.power(2, no_bits-1)
+    y = sum(b[0] * p) / scale
+
+    return y
 
 class DQNARQSolver():
     def __init__(self, just_replay=False, n_episodes=1000, max_env_steps=100, gamma=1.0, epsilon=1.0, epsilon_min=0.01, epsilon_log_decay=0.999999, alpha=0.01, alpha_decay=0.01, batch_size=256, quiet=False):
