+++ params/config.yaml
@@ -26,10 +26,10 @@ grgym_remote: # used if grgym_environment.run_local = False
     name: TX
     host: localhost # local/remote GnuRadio process
     rpc_port: 8080 # RPC port of local/remote GnuRadio process
-  node1:
-    name: RX
-    host: localhost2 # local/remote GnuRadio process
-    rpc_port: 8080 # RPC port of local/remote GnuRadio process
+#  node1:
+#    name: RX
+#    host: localhost2 # local/remote GnuRadio process
+#    rpc_port: 8080 # RPC port of local/remote GnuRadio process
   help: configuration used when gnuradio is running on remote machine, i.e. host name, Gnu Radio ID, Gnu Radio GRC filename, rpc port.
 
 grgym_scenario:
